# solutions feature
we're going to build the engine/solve feature. this is a new entity that's attached to a post which
has the  tag. new posts can be indicated as being a solution to the issue. so we're changing the models
as follows:

- post: gains a reply_to field
- solution: new model, has a post (A), has a user. indicates the posting user has solved the post  that post (A) was a reply to
- validation: new model, has a post, has a user. reply to a solution post. indicates a solution was validated, or refuted
- solutions can only be attached to posts that are not replies
- validations can only be a reply to a post which is has a solution

please generate me an ascii er digram to check this makes sense to you